fix: move transaction mode class; update schema on commit
