+++ src/harlequin/__init__.py
@@ -2,13 +2,15 @@ from harlequin.adapter import HarlequinAdapter, HarlequinConnection, HarlequinCu
 from harlequin.app import Harlequin
 from harlequin.autocomplete import HarlequinCompletion
 from harlequin.options import HarlequinAdapterOption, HarlequinCopyFormat
+from harlequin.transaction_mode import HarlequinTransactionMode
 
 __all__ = [
     "Harlequin",
     "HarlequinAdapter",
-    "HarlequinConnection",
-    "HarlequinCursor",
-    "HarlequinCopyFormat",
     "HarlequinAdapterOption",
     "HarlequinCompletion",
+    "HarlequinConnection",
+    "HarlequinCopyFormat",
+    "HarlequinCursor",
+    "HarlequinTransactionMode",
 ]
